refactor(engine): Share the permission watch between engines
The microphone permission watch (lazy open, sticky-state replay to late
subscribers, teardown on last unsubscribe) was implemented twice — inline in
WebSpeechEngine and again in demo/permissionWatch.ts — and the two `subscribe`
paths had already drifted apart. Extract a single createPermissionWatch into
src/, drive the built-in engine through it, and point the demo engines at it so
there is one implementation to maintain.
